Feature: Add tdi based treeshr hooks
This change will enable the implementation of tdi based
treeshr hooks which could be implemented in either tdi
or python. Like the C based hooks the tdi TreeShrHook() function
would be called with a action id, a treename, a shot number and a
node number (nid). A TreeShrHook.fun-template is included to
demonstrate how it is called during various tree operations.

The code will first try to call the C based implementation and if
there is no hooks library it will try the tdi based hooks.

This should implement the feature request:

https://github.com/MDSplus/mdsplus/issues/1349