Feature: Handle SIGINT in show server command
Typing <CTRL c> will interrupt the attempt to aquire the status of
an mdsip action server while executing the TCL "SHOW SERVER" command.
If a server system is not responding this will cancel the request to
that server and continue. It will not change the state of the action
server process.